move gameTest to PlayerEventTests
